Reorganising code and data file structure.
This commit mainly moves code and data files around in the directory
structure such that R libraries and python packages can be built from
the repository. All python code now lives in the "python" subdirectory,
and all R code in the "R" subdirectory. The data is in the
"inst/extdata" directory compatible with packaing data in R libraries
(the directory "extdata" will be insert one level above upon building
the library). For python build to distribute the R data folder as part
of the built posted package, it is moved into the "python/posted/"
directory prior to building via the "build_attach_data.py" script
specified in pyproject.toml.