Bugfix in generateTable.
This commit corrects the unit assumptions in the generation to
TEDataTables. Tables should all use the same unique reference dimension
and unit, such that the CAPEX reference unit has to be multiplied by 1a
(one year). This means, the CAPEX value in the TEDataTable is for a
production capacity in units of output per year.