Updated TEDataTable and calc routine col format.
This commit updates the format in which TEDataTables store their data
and how calc routines interact with it. Specifically, the calc function
of the TEDataTable class now has the 'keep' kwarg, which allows to
either keep all or only missing value columns in addition to calculated
(cost, ghgi, etc) columns.